Move directory of components
